Align command buffer flags

+++ src/hal/src/command/raw.rs
@@ -51,7 +51,7 @@ pub union ClearValueRaw {
 bitflags! {
     /// Option flags for various command buffer settings.
     #[derive(Default)]
-    pub struct CommandBufferFlags: u16 {
+    pub struct CommandBufferFlags: u32 {
         // TODO: Remove once 'const fn' is stabilized: https://github.com/rust-lang/rust/issues/24111
         /// No flags.
         const EMPTY = 0x0;
